# Partial sink pipelines should do something sensible with multiple calls