The extra work is necessary to avoid keeping garbage lying around. I did make an effort to avoid an extra array allocation for the common case of smaller partial pipelines (5 or less in the chain), though.