@isiahmeadows 2016-07-08
Still working on fixing this. Not ready to merge...