@dominictarr 2016-07-08
what problem does this solve?

I don't think this is necessary because you should never use a particular stream in two pipelines,
you should create new streams...