@dominictarr 2016-07-08
oh, do you mean garbage as in things that will be GC'd, so this saves creating an array and then throwing it out? is it possible to make a script/benchmark that shows how this is a problem?