@isiahmeadows 2016-07-08
@dominictarr This is solving two problems:

1. You can reuse a `pull()` sink. It might not always be a good idea, but it could be a thing.
2. The callback is more easily collectable.

The `switch` statement was actually just avoiding an additional array allocation for shorter things. If you want a benchmark, it's trivial to write, but I pulled it since it's unlikely to be a problem in practice - you'll have to do `reader(transform(readable))` by that point.