@dominictarr 2016-07-08
reusing a sink chain would definitely be at your own risk and not advised in general.
pull streams often have internal state (i.e. `highwatermark`, or `take`,
and these would have unintended behavior when reused.