@isiahmeadows 2016-07-08
The `read` callback may never get cleared otherwise until after the current tick ends, and it could cause a slow memory leak if you're not careful.