@dominictarr 2016-07-08
regards the GC thing though, it would be interesting to see _how much_ difference it does make, so a benchmark is appreciated. we have one here https://github.com/pull-stream/pull-stream/blob/master/benchmarks/pull.js from the last time someone made an optimization pull-request.