@isiahmeadows 2016-07-08
I benched three different methods for that:

1. The original `master`, without this patch.
2. Replacing the body of the returned function with `push.apply(null, [read].concat(args))`.
3. The `switch` statement I'm currently using in it.

Performance-wise, here's what I've found:

- `pull3` is about equal across all of them. I suspect the code path isn't hit.
- `pull_compose` is about 5-10% faster with the `switch`, but 5-10% slower with the `concat`. I'm pretty sure that returned function is hit in the second `pull` call of [this line](https://github.com/pull-stream/pull-stream/blob/master/benchmarks/pull.js#L35).
- `pull_chain` is roughly equal between the original and `concat`, but the `switch` version is slightly faster (not enough to mean much in practice).

Here's the results of 3 consecutive runs of each case:

Current `master`:

```
pull3*100000: 4073.384ms
pull_compose*100000: 4474.739ms
pull_chain*100000: 4008.915ms
pull3*100000: 4017.678ms
pull_compose*100000: 4432.934ms
pull_chain*100000: 3945.300ms
pull3*100000: 3983.170ms
pull_compose*100000: 4475.936ms
pull_chain*100000: 3952.188ms
```

This patch, using `concat`:

```
pull3*100000: 3902.641ms
pull_compose*100000: 4683.064ms
pull_chain*100000: 3979.576ms
pull3*100000: 4065.228ms
pull_compose*100000: 4761.607ms
pull_chain*100000: 4014.491ms
pull3*100000: 4187.816ms
pull_compose*100000: 4693.648ms
pull_chain*100000: 4020.804ms
```

This patch, using a `switch` statement:

```
pull3*100000: 4005.613ms
pull_compose*100000: 4108.665ms
pull_chain*100000: 3855.271ms
pull3*100000: 3972.441ms
pull_compose*100000: 4259.308ms
pull_chain*100000: 3931.466ms
pull3*100000: 3951.960ms
pull_compose*100000: 4152.017ms
pull_chain*100000: 3872.062ms
```

Here are the times, reformatted as a table:

Benchmark    | `master`      | This with `concat` | This with `switch`
-------------|---------------|--------------------|--------------------
pull3        | 1. 4073.384ms | 1. 3902.641ms      | 1. 4005.613ms
             | 2. 4017.678ms | 2. 4065.228ms      | 2. 3972.441ms
             | 3. 3983.170ms | 3. 4187.816ms      | 3. 3951.960ms
pull_compose | 1. 4474.739ms | 1. 4683.064ms      | 1. 4108.665ms
             | 2. 4432.934ms | 2. 4761.607ms      | 2. 4259.308ms
             | 3. 4475.936ms | 3. 4693.648ms      | 3. 4152.017ms
pull_chain   | 1. 4008.915ms | 1. 3979.576ms      | 1. 3855.271ms
             | 2. 3945.300ms | 2. 4014.491ms      | 2. 3931.466ms
             | 3. 3952.188ms | 3. 4020.804ms      | 3. 3872.062ms