@dominictarr 2016-07-08
looks good, but I think this can use https://github.com/pull-stream/pull-stream/pull/58/files#diff-2e88b737a946cc8fb8aacd25ff7ffe5aR15 `unshift` as in master instead of `concat`.

I just want to be completely clear that calling reusing the sink chain is not supported.
I'd sooner merge a thing that threw if called you `sink(read)` more than once.