@isiahmeadows 2016-07-08
@dominictarr I added that check and asserted it throws instead. I also changed the last branch (slow path) of the `switch` + the test to use `unshift` for consistency. Oh, and I also managed to speed it up further by using a local alias to the arguments. Here's the latest benchmark numbers (`node benchmark/pull.js` 3 times):

```
pull3*100000: 3891.233ms
pull_compose*100000: 4133.924ms
pull_chain*100000: 3867.556ms
pull3*100000: 3890.908ms
pull_compose*100000: 4076.031ms
pull_chain*100000: 3866.232ms
pull3*100000: 3818.921ms
pull_compose*100000: 4040.969ms
pull_chain*100000: 3815.148ms
```

In a table, compared to `master`:

Benchmark    | `master`      | Current
-------------|---------------|---------------
pull3        | 1. 4073.384ms | 1. 3891.233ms
             | 2. 4017.678ms | 2. 3890.908ms
             | 3. 3983.170ms | 3. 3818.921ms
pull_compose | 1. 4474.739ms | 1. 4133.924ms
             | 2. 4432.934ms | 2. 4076.031ms
             | 3. 4475.936ms | 3. 4040.969ms
pull_chain   | 1. 4008.915ms | 1. 3867.556ms
             | 2. 3945.300ms | 2. 3866.232ms
             | 3. 3952.188ms | 3. 3815.148ms