@dominictarr 2016-08-31
sorry i let this slip, merged into 3.4.4